change weights to float

+++ src/main/java/org/agilemethod/pair10/grade_system/Grade.java
@@ -8,7 +8,7 @@ public class Grade {
 
     }
 
-    public int calculateTotalGrade(int[] weights) {
+    public int calculateTotalGrade(float[] weights) {
         return 0;
     }
 }
